🪝useLoading Hook
To access the loading state in our app or set a new loading state, try  the useLoading hook 👉🏿
[loading, setLoading] = useLoading( )

I have also moved the ProgressBar component to the Layout component so that the bar can show above the NavBar component - the loading state in our app has also been passed to the bar. Conclusively, the loading state is Global - a redux state and available anywhere within the app.
To prevent a continuous loading state, the loading state will automatically set to false after 10 seconds.